+++ priv/repo/migrations/20200207111346_add_domain_external_id.exs
@@ -1,0 +1,9 @@
+defmodule TdBg.Repo.Migrations.AddDomainExternalId do
+  use Ecto.Migration
+
+  def change do
+    alter table(:domains) do
+      add :external_id, :string, null: true
+    end
+  end
+end